+++ sample-service/src/main/java/uk/ac/ox/ndph/mts/sampleservice/DemoApp.java
@@ -1,6 +1,6 @@
-package com.microsoft.baseapp;
+package uk.ac.ox.ndph.mts.sampleservice;
 
-import com.microsoft.baseapp.config.IConfigService;
+import uk.ac.ox.ndph.mts.sampleservice.config.IConfigService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
