+++ sample-service/src/main/java/uk/ac/ox/ndph/mts/sampleservice/config/IConfigService.java
@@ -1,4 +1,4 @@
-package com.microsoft.baseapp.config;
+package uk.ac.ox.ndph.mts.sampleservice.config;
 
 public interface IConfigService {
     String getSecret();
